Improve testability of Interactions
