+++ TF_Service_dotNet/TouchFree_Service/Startup.cs
@@ -17,20 +17,7 @@ namespace Ultraleap.TouchFree.Service
         // For more information on how to configure your application, visit https://go.microsoft.com/fwlink/?LinkID=398940
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddUpdateBehaviour();
-            services.AddConfig();
-
-            services.AddTrackingConnectionManager();
-            services.AddTrackingDiagnosticApi();
-            services.AddHandManager();
-            services.AddVirtualScreen();
-
-            services.AddPositioning();
-
-            services.AddClientConnectionManager();
-            services.AddMessageQueueHandlers();
-
-            services.AddInteractions();
+            services.ConfigureTouchFreeServices();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
